refactor(import): use dynamic import for subcommand
