+++ packages/opencode/src/index.ts
@@ -21,7 +21,7 @@ import { StatsCommand } from "./cli/cmd/stats"
 import { McpCommand } from "./cli/cmd/mcp"
 import { GithubCommand } from "./cli/cmd/github/command"
 import { ExportCommand } from "./cli/cmd/export/command"
-import { ImportCommand } from "./cli/cmd/import"
+import { ImportCommand } from "./cli/cmd/import/command"
 import { AttachCommand } from "./cli/cmd/tui/attach"
 import { TuiThreadCommand } from "./cli/cmd/tui/thread"
 import { AcpCommand } from "./cli/cmd/acp/command"
